+++ tests/test_WikiData.py
@@ -11,6 +11,7 @@ class TestWikiData(unittest.TestCase):
     ''' test the WikiData proceedings titles source '''
 
     def setUp(self):
+        self.forceCaching=False
         pass
 
     def tearDown(self):
@@ -18,16 +19,21 @@ class TestWikiData(unittest.TestCase):
 
     def testWikiData(self):
         ''' test getting the wikidata information '''
-        wd=WikiData(debug=True)
-        cacheFile=wd.em.getCacheFile()
-        if os.path.isfile(cacheFile):
-            os.remove(cacheFile)
-        wd.cacheEvents()
+        wd=WikiData(debug=False,profile=True)
+        if self.forceCaching:
+            cacheFile=wd.em.getCacheFile()
+            if os.path.isfile(cacheFile):
+                os.remove(cacheFile)
+        if not wd.em.isCached():        
+            wd.cacheEvents(limit=1000000,batchSize=500)
+        else:
+            wd.em.fromStore()    
         print("found %d wikidata events" % (len(wd.em.events)))
         self.assertTrue(len(wd.em.events)>310)
-        size=os.stat(cacheFile).st_size
-        print (size)
-        self.assertTrue(size>208000)
+        if wd.em.mode=='json':
+            size=os.stat(cacheFile).st_size
+            print (size)
+            self.assertTrue(size>208000)
         pass
 
 if __name__ == "__main__":
